+++ engine/runtime/src/bench/scala/org/enso/interpreter/bench/fixtures/semantic/AtomFixtures.scala
@@ -20,7 +20,7 @@ class AtomFixtures extends DefaultInterpreterRunner {
   val millionElementList = buildInputList(million)
 
   val generateListCode =
-    """from Builtins import all
+    """from Standard.Builtins import all
       |
       |main = length ->
       |    generator = acc -> i -> if i == 0 then acc else @Tail_Call generator (Cons i acc) (i - 1)
@@ -31,7 +31,7 @@ class AtomFixtures extends DefaultInterpreterRunner {
   val generateList = getMain(generateListCode)
 
   val generateListQualifiedCode =
-    """from Builtins import all
+    """from Standard.Builtins import all
       |
       |main = length ->
       |    generator = acc -> i -> if i == 0 then acc else @Tail_Call generator (Builtins.cons i acc) (i - 1)
@@ -42,7 +42,7 @@ class AtomFixtures extends DefaultInterpreterRunner {
   val generateListQualified = getMain(generateListQualifiedCode)
 
   val reverseListCode =
-    """from Builtins import all
+    """from Standard.Builtins import all
       |
       |main = list ->
       |    reverser = acc -> list -> case list of
@@ -55,7 +55,7 @@ class AtomFixtures extends DefaultInterpreterRunner {
   val reverseList = getMain(reverseListCode)
 
   val reverseListMethodsCode =
-    """from Builtins import all
+    """from Standard.Builtins import all
       |
       |Cons.reverse = acc -> case this of
       |    Cons h t -> @Tail_Call t.reverse (Cons h acc)
@@ -69,7 +69,7 @@ class AtomFixtures extends DefaultInterpreterRunner {
   val reverseListMethods = getMain(reverseListMethodsCode)
 
   val sumListCode =
-    """from Builtins import all
+    """from Standard.Builtins import all
       |
       |main = list ->
       |    summator = acc -> list -> case list of
@@ -82,7 +82,7 @@ class AtomFixtures extends DefaultInterpreterRunner {
   val sumList = getMain(sumListCode)
 
   val sumListLeftFoldCode =
-    """from Builtins import all
+    """from Standard.Builtins import all
       |
       |main = list ->
       |    fold = f -> acc -> list -> case list of
@@ -95,7 +95,7 @@ class AtomFixtures extends DefaultInterpreterRunner {
   val sumListLeftFold = getMain(sumListLeftFoldCode)
 
   val sumListFallbackCode =
-    """from Builtins import all
+    """from Standard.Builtins import all
       |
       |main = list ->
       |    summator = acc -> list -> case list of
@@ -108,7 +108,7 @@ class AtomFixtures extends DefaultInterpreterRunner {
   val sumListFallback = getMain(sumListFallbackCode)
 
   val sumListMethodsCode =
-    """from Builtins import all
+    """from Standard.Builtins import all
       |
       |Nil.sum = acc -> acc
       |Cons.sum = acc -> case this of
@@ -121,7 +121,7 @@ class AtomFixtures extends DefaultInterpreterRunner {
   val sumListMethods = getMain(sumListMethodsCode)
 
   val mapReverseListCode =
-    """from Builtins import all
+    """from Standard.Builtins import all
       |
       |Nil.mapReverse = f -> acc -> acc
       |Cons.mapReverse = f -> acc -> case this of
@@ -134,7 +134,7 @@ class AtomFixtures extends DefaultInterpreterRunner {
   val mapReverseList = getMain(mapReverseListCode)
 
   val mapReverseListCurryCode =
-    """from Builtins import all
+    """from Standard.Builtins import all
       |
       |Nil.mapReverse = f -> acc -> acc
       |Cons.mapReverse = f -> acc -> case this of