+++ engine/runtime/src/main/scala/org/enso/compiler/pass/resolve/DocumentationComments.scala
@@ -147,7 +147,7 @@ case object DocumentationComments extends IRPass {
       case _: IR.Name.Annotation =>
         throw new CompilerError(
           "Annotations should already be associated by the point of " +
-            "documentation comment resolution."
+          "documentation comment resolution."
         )
     }
 
@@ -168,14 +168,14 @@ case object DocumentationComments extends IRPass {
     */
   private def resolveIr(ir: IR): IR =
     ir match {
-      case module: IR.Module              => resolveModule(module)
-      case expr: IR.Expression            => resolveExpression(expr)
-      case df: IR.Module.Scope.Definition => resolveDefinition(df)
-      case imp: IR.Module.Scope.Import    => imp
-      case exp: IR.Module.Scope.Export    => exp
-      case arg: IR.CallArgument           => arg
-      case arg: IR.DefinitionArgument     => arg
-      case pat: IR.Pattern                => pat
+      case module: IR.Module                  => resolveModule(module)
+      case expr: IR.Expression                => resolveExpression(expr)
+      case df: IR.Module.Scope.Definition     => resolveDefinition(df)
+      case imp: IR.Module.Scope.Import        => imp
+      case exp: IR.Module.Scope.Export.Module => exp
+      case arg: IR.CallArgument               => arg
+      case arg: IR.DefinitionArgument         => arg
+      case pat: IR.Pattern                    => pat
     }
 
   // === Metadata =============================================================