+++ engine/runtime/src/test/scala/org/enso/compiler/test/CompilerTest.scala
@@ -244,7 +244,7 @@ trait CompilerRunner {
     compilerConfig: CompilerConfig               = defaultConfig
   ): ModuleContext = {
     ModuleContext(
-      module            = Module.empty(moduleName),
+      module            = Module.empty(moduleName, null),
       freshNameSupply   = freshNameSupply,
       passConfiguration = passConfiguration,
       compilerConfig    = compilerConfig
@@ -267,7 +267,7 @@ trait CompilerRunner {
     passConfiguration: Option[PassConfiguration] = None,
     compilerConfig: CompilerConfig               = defaultConfig
   ): InlineContext = {
-    val mod = Module.empty(QualifiedName.simpleName("Test_Module"))
+    val mod = Module.empty(QualifiedName.simpleName("Test_Module"), null)
     mod.unsafeBuildIrStub()
     InlineContext(
       module            = mod,