+++ engine/runtime/src/test/scala/org/enso/interpreter/test/semantic/LambdaChainingTest.scala
@@ -56,7 +56,7 @@ class LambdaChainingTest extends InterpreterTest {
 
     "work properly with lazy parameters" in {
       val code =
-        """from Builtins import all
+        """from Standard.Builtins import all
           |
           |main =
           |    fn = a -> ~b -> ~c ->