+++ engine/runtime/src/test/scala/org/enso/std/test/NumberTest.scala
@@ -10,7 +10,7 @@ class NumberTest extends InterpreterTest {
   ): Unit = {
     "support equality comparisons" in {
       val code =
-        """from Builtins import all
+        """from Standard.Builtins import all
           |
           |main =
           |    IO.println 7==5